# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vault API
  version: 1.0.0
  description: |
    Zero-knowledge password vault backend. The server stores encrypted blobs and metadata only;
    clients derive encryption keys locally and encrypt vault contents before upload.

    ## Authentication

    Protected endpoints require a Bearer JWT access token obtained from login, refresh, or
    recovery verify. Include the header:

    `Authorization: Bearer <access_token>`

    ## Encrypted payloads

    `encrypted_data` and `encrypted_item_key` fields are raw bytes serialized as base64 strings
    in JSON.

    ## Response naming

    Vault domain objects returned directly from vault CRUD endpoints use Go's default JSON field
    names (PascalCase, e.g. `ID`, `UserID`). Handler-wrapped responses use snake_case
    (e.g. `access_token`, `vault_item_id`).

  contact:
    name: Vault API
  license:
    name: MIT
servers:
  - url: http://localhost:8081
    description: Local development (default PORT=8081)
...
paths:
...
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
        "400":
          $ref: "#/components/responses/BadRequestText"
        "401":
          $ref: "#/components/responses/UnauthorizedText"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalErrorText"
...
  /api/v1/mfa/enable:
    post:
      tags: [MFA]
      summary: Begin MFA enrollment
      operationId: enableMFA
      description: |
        Requires the account password. Generates a TOTP secret. MFA is not active until
        `/mfa/verify` succeeds with a valid code. The secret is encrypted at rest.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MFAEnableRequest"
      responses:
        "200":
          description: TOTP setup details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MFASetupResponse"
        "400":
          $ref: "#/components/responses/BadRequestText"
        "401":
          $ref: "#/components/responses/UnauthorizedText"
        "409":
          description: MFA already enabled
          content:
            text/plain:
              schema:
                type: string
                example: mfa already enabled
        "500":
          $ref: "#/components/responses/InternalErrorText"
  /api/v1/mfa/verify:
    post:
      tags: [MFA]
      summary: Confirm MFA enrollment
      operationId: verifyMFA
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TOTPCodeRequest"
      responses:
        "204":
          description: MFA enabled
        "400":
          description: MFA not enabled (no pending secret) or invalid JSON
          content:
            text/plain:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/UnauthorizedText"
        "409":
          description: MFA already enabled
          content:
            text/plain:
              schema:
                type: string
                example: mfa already enabled
        "500":
          $ref: "#/components/responses/InternalErrorText"
  /api/v1/mfa/disable:
    post:
      tags: [MFA]
      summary: Disable MFA
      operationId: disableMFA
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TOTPCodeRequest"
      responses:
        "204":
          description: MFA disabled
        "400":
          description: MFA not enabled
          content:
            text/plain:
              schema:
                type: string
                example: mfa not enabled
        "401":
          $ref: "#/components/responses/UnauthorizedText"
        "500":
          $ref: "#/components/responses/InternalErrorText"
  /api/v1/recovery/generate:
    post:
      tags: [Recovery]
      summary: Generate recovery codes
      operationId: generateRecoveryCodes
      description: |
        Requires MFA to be enabled, plus account password and a current TOTP code.
        Replaces any existing unused codes.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryGenerateRequest"
      responses:
        "200":
          description: Plaintext recovery codes (shown once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryCodesResponse"
        "400":
          description: MFA must be enabled first or invalid JSON
          content:
            text/plain:
              schema:
                type: string
                example: mfa must be enabled before generating recovery codes
        "401":
          $ref: "#/components/responses/UnauthorizedText"
        "500":
          $ref: "#/components/responses/InternalErrorText"
  /api/v1/recovery/verify:
    post:
      tags: [Recovery]
      summary: Login with recovery code
      operationId: verifyRecovery
      description: |
        Bypasses MFA when a valid one-time recovery code is supplied. On success the server
        disables MFA (forcing re-enrollment), deletes remaining recovery codes, and revokes
        other sessions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryVerifyRequest"
      responses:
        "200":
          description: Login successful; recovery code consumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPairResponse"
        "400":
          $ref: "#/components/responses/BadRequestText"
        "401":
          description: Invalid credentials or recovery code
          content:
            text/plain:
              schema:
                type: string
        "500":
          $ref: "#/components/responses/InternalErrorText"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from login, refresh, or recovery verify
...
  responses:
    BadRequestText:
      description: Invalid request
      content:
        text/plain:
          schema:
            type: string
    UnauthorizedText:
      description: Missing or invalid access token
      content:
        text/plain:
          schema:
            type: string
            example: unauthorized
...
    InternalErrorText:
      description: Internal server error
      content:
        text/plain:
          schema:
            type: string
    TooManyRequests:
      description: Rate limit exceeded (auth endpoints)
      content:
        text/plain:
          schema:
            type: string
  schemas:
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...
    TokenPairResponse:
      type: object
      required: [access_token, refresh_token]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
    AccessTokenResponse:
      type: object
      required: [access_token]
      properties:
        access_token:
          type: string
...
    MFASetupResponse:
      type: object
      required: [secret, otpauth_url]
      properties:
        secret:
          type: string
          description: Base32 TOTP secret (store securely until verified)
        otpauth_url:
          type: string
          format: uri
          description: otpauth:// URI for authenticator apps
    TOTPCodeRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          description: 6-digit TOTP code
    MFAEnableRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          format: password
          description: Current account password (step-up authentication)
    RecoveryGenerateRequest:
      type: object
      required: [password, code]
      properties:
        password:
          type: string
          format: password
          description: Current account password
        code:
          type: string
          description: Current 6-digit TOTP code
    RecoveryCodesResponse:
      type: object
      required: [recovery_codes]
      properties:
        recovery_codes:
          type: array
          items:
            type: string
          minItems: 10
          maxItems: 10
          description: Plaintext one-time codes; not retrievable after this response
    RecoveryVerifyRequest:
      type: object
      required: [email, password, recovery_code]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        recovery_code:
          type: string
        device_name:
          type: string
...